requirements.txt
Flask==1.1.2
google-cloud-firestore==2.0.2
gunicorn==20.0.4
pytest==6.2.1